backend-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
reportlab==4.2.2
email-validator==2.2.0
pinecone-client==5.0.1
google-generativeai==0.8.3
httpx==0.27.2